Counting DNA nucleotides and figuring out github
This was part of my second (60 min) work session. As you can see I am still figuring this out because I didn't enter in my journal entry for my first 45 minutes. The code is having issues but this is how far I've gotten in the hour. Plus I used some of this work session to try and figure out github. 
Here is my journal entry from session one:
I was learning how to use Github and also used a lot of my time researching which project to do. I spent a lot of my time relearning bioinformatics and how to approach the first problem.