finishing Counting DNA nucleotides, DNA to RNA
January 17 (6:30-8:30)
first hour
Writing this journal entry as I go. I'm trying to redo my dnaCount function because it is not working. I pasted in my second try and commented out my first try. The second try, I'm 30 minutes in and I made it so that the code returns my DNA sequence (lol not supposed to happen). (false alarm I fixed it but the output is still (209, 241, 210, 239) so) I'm going to scratch this out but still, show it in the comments. And I'm going to try and have an extremely different approach to see if it works better. (another false alarm I actually just used the wrong data) sorry, this is so messy. What I thought was wrong was actually right it was just different from the sample datasets because I used the DNA files. The second try was a success(I spelled my own function wrong oops)! During this first hour, I also finished the DNA to RNA problem. It was pretty easy and quick as I had the format of the first problem to help me. \
Second Hour
I Started the REVC but I didn't finish. I keep getting the "KeyError: '\n'" but I have no idea what it means; next work session I'll research it more to understand the error message and fix my code. As I am working on all of these problems it is getting easier and easier for me to write functions and loops. I pasted in the progress I got on REVC.