Fixing REVC and starting PROT
1/18/22 - 8:00pm - 9:00pm
the key error is when you attempt to access an item in the dictionary that does not exist (Stack Overflow). update in the hour I was not able to fix the revc but I was able to start prot. I have no idea what's going wrong with my revc but I will paste it in and show you anyways. This file has all the code for this project but I should have more organized uploaded files within my project folder. (if it's uploaded there that is the final code that works). Questions will be asked at office hours :D